value stmt: replaced integer by long; arithmetic values

+++ src/main/java/io/proleap/cobol/asg/metamodel/IntegerLiteral.java
@@ -16,5 +16,5 @@ public interface IntegerLiteral extends CobolDivisionElement, TypedElement {
 	@Override
 	IntegerLiteralContext getCtx();
 
-	Integer getValue();
+	Long getValue();
 }
